properly render Rat Master armor
