+++ gradle.properties
@@ -10,7 +10,7 @@ archives_base_name = ratsmischief
 minecraft_version = 1.20.1
 yarn_mappings = 10
 
-loader_version = 0.15.11
+loader_version = 0.16.14
 
 # Fabric API
 fabric_api_version = 0.92.6
