--- src/main/java/ladysnake/ratsmischief/common/util/EntityRendererWrapper.java
@@ -1,7 +1,0 @@
-package ladysnake.ratsmischief.common.util;
-
-import net.minecraft.client.render.entity.EntityRendererFactory;
-
-public interface EntityRendererWrapper {
-	EntityRendererFactory.Context mischief$getContext();
-}
