--- src/main/java/ladysnake/ratsmischief/common/util/PlayerEntityRendererWrapper.java
@@ -1,5 +1,0 @@
-package ladysnake.ratsmischief.common.util;
-
-public interface PlayerEntityRendererWrapper {
-	boolean mischief$isSlim();
-}
