+++ src/main/java/ladysnake/ratsmischief/mixin/client/PlayerEntityRendererMixin.java
@@ -2,7 +2,6 @@ package ladysnake.ratsmischief.mixin.client;
 
 import com.llamalad7.mixinextras.sugar.Local;
 import ladysnake.ratsmischief.common.init.ModItems;
-import ladysnake.ratsmischief.common.util.PlayerEntityRendererWrapper;
 import net.minecraft.client.network.AbstractClientPlayerEntity;
 import net.minecraft.client.render.entity.EntityRendererFactory;
 import net.minecraft.client.render.entity.LivingEntityRenderer;
@@ -13,17 +12,13 @@ import net.minecraft.item.ItemStack;
 import net.minecraft.util.Arm;
 import net.minecraft.util.Hand;
 import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 @Mixin(PlayerEntityRenderer.class)
-public abstract class PlayerEntityRendererMixin extends LivingEntityRenderer<AbstractClientPlayerEntity, PlayerEntityModel<AbstractClientPlayerEntity>> implements PlayerEntityRendererWrapper {
-	@Unique
-	private boolean isSlim = true;
-
+public abstract class PlayerEntityRendererMixin extends LivingEntityRenderer<AbstractClientPlayerEntity, PlayerEntityModel<AbstractClientPlayerEntity>> {
 	public PlayerEntityRendererMixin(EntityRendererFactory.Context ctx, PlayerEntityModel<AbstractClientPlayerEntity> model, float shadowRadius) {
 		super(ctx, model, shadowRadius);
 	}
@@ -34,11 +29,6 @@ public abstract class PlayerEntityRendererMixin extends LivingEntityRenderer<Abs
 		if (stack.isOf(ModItems.RAT)) {
 			cir.setReturnValue(BipedEntityModel.ArmPose.BOW_AND_ARROW);
 		}
-	}
-
-	@Inject(method = "<init>", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/render/entity/PlayerEntityRenderer;addFeature(Lnet/minecraft/client/render/entity/feature/FeatureRenderer;)Z", ordinal = 0))
-	private void mischief$masterArmor(EntityRendererFactory.Context ctx, boolean slim, CallbackInfo ci) {
-		this.isSlim = slim;
 	}
 
 	@Inject(method = "setModelPose", at = @At(value = "TAIL"))
@@ -54,10 +44,5 @@ public abstract class PlayerEntityRendererMixin extends LivingEntityRenderer<Abs
 				if (offHandRat) playerEntityModel.rightArmPose = BipedEntityModel.ArmPose.BOW_AND_ARROW;
 			}
 		}
-	}
-
-	@Override
-	public boolean mischief$isSlim() {
-		return this.isSlim;
 	}
 }
