+++ src/main/resources/ratsmischief.mixins.json
@@ -15,7 +15,6 @@
     "WardenEntityMixin"
   ],
   "client": [
-    "client.ArmorFeatureRendererMixin",
     "client.BipedEntityModelMixin",
     "client.EntityRendererMixin",
     "client.HeadFeatureRendererMixin",
